Moved middleware into service
